Rewrite

+++ src/object/object.cpp
@@ -50,7 +50,7 @@ ObjString* copyString(const char* chars, size_t length) {
     return allocateString(heapChars, length, hash);
 }
 
-ObjString* takeString(char* chars, int length) {
+ObjString* takeString(char* chars, size_t length) {
     uint32_t hash = hashString(chars, length);
 
     ObjString* interned = HashTable::findString(&vm.strings, chars, length, hash);
@@ -74,7 +74,7 @@ void concatenate() {
     ObjString* b = AS_STRING(pop());
     ObjString* a = AS_STRING(pop());
 
-    int length = a->length + b->length;
+    size_t length = a->length + b->length;
     char* chars = Memory::ALLOCATE<char>(length + 1);
     memcpy(chars, a->chars, a->length);
     memcpy(chars + a->length, b->chars, b->length);
