+++ src/object/object.hpp
@@ -14,13 +14,13 @@ struct Obj {
 
 struct  ObjString {
     Obj obj;
-    int length;
+    size_t length;
     char* chars;
     uint32_t hash;
 };
 
 ObjString* copyString(const char* chars, size_t length);
-ObjString* takeString(char* chars, int length);
+ObjString* takeString(char* chars, size_t length);
 
 void printObject(Value value);
 void concatenate();
